Handle renamed repos
Properly deduplicate repos that were renamed on GitHub. This
causes a redirect from the old repo to the new repo, which
broke de-duplication.

Adds a unit test for this scenario, and fixes up some
of the other tests.

Also fixes the authentication test, which broke when using an
OAuth app token rather than a personal access token for GitHub auth.