Change approach to identifying repos, remove all features except SUSHI
This now relies entirely on the GitHub Search API for identifying repos
with FSH, vs. checking manually for the existence of .fsh files in
the correct location.

It also stops checking every single branch in a repo, and removes checking
of use of FSH language features (profiles, value sets, etc.) as most
repos are using all these features.

This gets the runtime down to about 7 minutes for an update. The length
is primarily due to API rate limiting, and can't really be worked around
without adding much more aggressive/complex caching.